Split Oscilloscope.h
